Adding a controller-wide title definition
If the content_for is not defined, then we look for a `meta.title.controller_name` translation.
If no key is defined, `RDV Solidarités` is kept. The spec on this is would be either moving lots of files, either meta-ish and I don't much like that.

Don't hesitate to propose way to spec this, I would be glad to hear it :)

Refers to #1020