Encode the text/calendar part of emails as base64 too
This is a return to what we used to do before #1531.

Quoted-printable would make sense, and this is what Google Calendar uses.
However, with quoted-printable, Rails (in quoted-printable.rb) uses “pack("M")” which gives `=OD\n`.
I’m not 100% sure if this is fine by the specs.(See RFC2045) But:
* We now that Zimbra doesn’t like it very much: when the application/ics attachment is encoded as such, zimbra decodes it as CRCRLF and its icalendar parsing library crashes.
* Outlook apparently does not automatically interpret the text/calendar part anymore, since #1531