Fix some json serialization weirdness in ruby 3
This broke:
 * the feature test:
   * Admin can configure the organisation CRUD on motifs
 * json encoding of DateTime objects via jbuilder
   * but not via blueprints.

For some reason in Ruby 3, rendering ActiveRecord objects takes a different path.
I tracked it down to ActiveSupport::ToJsonWithActiveSupportEncoder:
In ruby 3, options.is_a?(::JSON::State), in ruby 2.7 it uses ActiveSupport::JSON regularly.

The best source I can find is this SO discussion: https://stackoverflow.com/questions/66871265/to-json-on-activerecord-object-ruby-3-0

The workaround is to manually convert to json.