🧹 Index Absence.expired_cached
In general, it's not useful to index boolean column in PG, because it's actually faster to scan the whole DB sequentially than perform I/O to fetch the matching rows. However, it becomes relevant if there are few rows to return. When looking at production data, PlageOuverture.expired.explain uses Seq Scan, but PlageOuverture.not_expired.explain uses Index Scan. For Absences, I expect the ratio of not_expired to expired to be even lower, so the index will certainly be useful.

🧹 We’ll need to run UpdateExpirationsJob once before it becomes useful.