Fix an issue with rdv.with_agents being used with new rdvs
We’re calling `Rdv.with_agent(rdv.agents)`. If `rdv` is a new_record?, the `rdv.agents` can’t actually be used to build ActiveRecord query, since the id of the joined rdv is nil. In that case, the composed query would return no results. Let’s just force it to an array in that case.